start first controller todo allow origins

+++ src/main/java/com/eva/eva/repositoryes/SkillRepository.java
@@ -4,6 +4,6 @@ import com.eva.eva.entities.Student;
 import com.eva.eva.entities.skill.Skill;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface SkillRepository extends PagingAndSortingRepository<Skill, Long> {
+public interface SkillRepository extends PagingAndSortingRepository<Skill, Integer> {
 
 }
