+++ src/main/java/com/eva/eva/repositoryes/StudentRepository.java
@@ -3,6 +3,6 @@ package com.eva.eva.repositoryes;
 import com.eva.eva.entities.Student;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface StudentRepository extends PagingAndSortingRepository<Student, Long> {
+public interface StudentRepository extends PagingAndSortingRepository<Student, Integer> {
 
 }
